Collected methods in manipulate_branch and update method names after merge. May consider placing common methods in xxx_operations or common.py for later
